Restructure project to organize models under pkg
Relocated various model files to the pkg directory for better organization and maintainability. Also moved `api_wallet.go` into the internal/websocket directory and updated import paths accordingly.
