+++ pkg/models/book_summary.go
